Repurpose the background view from the SegmentedControlDemo to update the PillButtonBarDemo also
