+++ ios/FluentUI.Demo/FluentUI.Demo/Demos/SegmentedControlDemoController.swift
@@ -79,7 +79,15 @@ class SegmentedControlDemoController: DemoController {
                               action: #selector(updateLabel(forControl:)),
                               for: .valueChanged)
 
-        let backgroundView = BackgroundView(style: style)
+        let backgroundStyle: ColoredBackgroundStyle = {
+            switch style {
+            case .primaryPill:
+                return .neutral
+            case .onBrandPill:
+                return .brand
+            }
+        }()
+        let backgroundView = ColoredBackgroundView(style: backgroundStyle)
         segmentedControls.append(pillControl)
 
         backgroundView.translatesAutoresizingMaskIntoConstraints = false
@@ -145,57 +153,4 @@ extension SegmentedControlDemoController: DemoAppearanceDelegate {
             return FontInfo(name: "Papyrus", size: 10.0, weight: .regular)
         }
     }
-}
-
-class BackgroundView: UIView {
-    init(style: SegmentedControlStyle) {
-        self.style = style
-        super.init(frame: .zero)
-        updateBackgroundColor()
-
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(themeDidChange),
-                                               name: .didChangeTheme,
-                                               object: nil)
-    }
-
-    required init?(coder: NSCoder) {
-        preconditionFailure("init(coder:) has not been implemented")
-    }
-
-    override func didMoveToWindow() {
-        super.didMoveToWindow()
-        updateBackgroundColor()
-    }
-
-    @objc func themeDidChange() {
-        updateBackgroundColor()
-    }
-
-    func updateBackgroundColor() {
-        let lightColor: UIColor
-        let darkColor: UIColor
-        switch style {
-        case .primaryPill:
-            if let fluentTheme = window?.fluentTheme {
-                lightColor = UIColor(dynamicColor: fluentTheme.aliasTokens.backgroundColors[.neutral1])
-                darkColor = UIColor(dynamicColor: fluentTheme.aliasTokens.backgroundColors[.neutral4])
-            } else {
-                lightColor = Colors.navigationBarBackground
-                darkColor = Colors.navigationBarBackground
-            }
-        case .onBrandPill:
-            if let fluentTheme = window?.fluentTheme {
-                lightColor = UIColor(dynamicColor: fluentTheme.aliasTokens.backgroundColors[.brandRest])
-                darkColor = UIColor(dynamicColor: fluentTheme.aliasTokens.backgroundColors[.neutral4])
-            } else {
-                lightColor = Colors.communicationBlue
-                darkColor = Colors.navigationBarBackground
-            }
-        }
-
-        backgroundColor = UIColor(light: lightColor, dark: darkColor)
-    }
-
-    let style: SegmentedControlStyle
 }
